fix usage of temperature filter
+ add debugging

+++ Software/Core/Inc/bmu_window_filter.h
@@ -15,16 +15,16 @@
 #define BMU_MAX_FILTER_SIZE 8
 
 typedef struct bmu_window_filter {
-	uint32_t current_filtered;
-	uint32_t prev_values[BMU_MAX_FILTER_SIZE];
+	uint16_t current_filtered;
+	uint16_t prev_values[BMU_MAX_FILTER_SIZE];
 	uint32_t running_sum;
-	uint32_t current_idx;
+	uint16_t current_idx;
 	bool initialized;
 	uint16_t window_size;
 } bmu_window_filter_t;
 
-void bmu_window_filter_initialize(bmu_window_filter_t *filter, uint32_t initial_value, uint16_t window_size);
+void bmu_window_filter_initialize(bmu_window_filter_t *filter, uint16_t initial_value, uint16_t window_size);
 
-void bmu_window_filter_update(bmu_window_filter_t *filter, uint32_t new_value);
+void bmu_window_filter_update(bmu_window_filter_t *filter, uint16_t new_value);
 
 #endif /* INC_BMU_WINDOW_FILTER_H_ */
